+++ sealed_annotations/example/result_complex.sealed.dart
@@ -77,7 +77,7 @@ abstract class Result<D extends num, E extends Object> {
     }
   }
 
-  R whenOrElse<R extends Object?>({
+  R maybeWhen<R extends Object?>({
     R Function(ResultSuccess<D, E> success)? success,
     R Function(ResultError<D, E> error)? error,
     R Function(ResultMixed<D, E> mixed)? mixed,
@@ -95,7 +95,7 @@ abstract class Result<D extends num, E extends Object> {
     }
   }
 
-  void whenPartial({
+  void partialWhen({
     void Function(ResultSuccess<D, E> success)? success,
     void Function(ResultError<D, E> error)? error,
     void Function(ResultMixed<D, E> mixed)? mixed,