+++ sealed_writer/lib/src/writer/top/match/top_match_writer.dart
@@ -1,8 +1,8 @@
 import 'package:meta/meta.dart';
 import 'package:sealed_writer/src/manifest/manifest.dart';
 import 'package:sealed_writer/src/writer/base/base_writer.dart';
-import 'package:sealed_writer/src/writer/top/match/top_match_when_or_else_writer.dart';
-import 'package:sealed_writer/src/writer/top/match/top_match_when_partial_writer.dart';
+import 'package:sealed_writer/src/writer/top/match/top_match_maybe_when_writer.dart';
+import 'package:sealed_writer/src/writer/top/match/top_match_partial_when_writer.dart';
 import 'package:sealed_writer/src/writer/top/match/top_match_when_writer.dart';
 
 /// match method writer
@@ -11,19 +11,19 @@ import 'package:sealed_writer/src/writer/top/match/top_match_when_writer.dart';
 class TopMatchWriter extends BaseWriter {
   TopMatchWriter(Manifest manifest)
       : topMatchWhenWriter = TopMatchWhenWriter(manifest),
-        topMatchWhenOrElseWriter = TopMatchWhenOrElseWriter(manifest),
-        topMatchWhenPartialWriter = TopMatchWhenPartialWriter(manifest),
+        topMatchMaybeWhenWriter = TopMatchMaybeWhenWriter(manifest),
+        topMatchPartialWhenWriter = TopMatchPartialWhenWriter(manifest),
         super(manifest);
 
   final TopMatchWhenWriter topMatchWhenWriter;
 
-  final TopMatchWhenOrElseWriter topMatchWhenOrElseWriter;
+  final TopMatchMaybeWhenWriter topMatchMaybeWhenWriter;
 
-  final TopMatchWhenPartialWriter topMatchWhenPartialWriter;
+  final TopMatchPartialWhenWriter topMatchPartialWhenWriter;
 
   Iterable<String> topMatchMethods() => [
         topMatchWhenWriter.topMatchWhen(),
-        topMatchWhenOrElseWriter.topMatchWhenOrElse(),
-        topMatchWhenPartialWriter.topMatchWhenPartial(),
+        topMatchMaybeWhenWriter.topMatchMaybeWhen(),
+        topMatchPartialWhenWriter.topMatchPartialWhen(),
       ];
 }